# ads_tiktok/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/sessions":{"post":{"operationId":"SessionController_create","parameters":[{"name":"x-api-key","required":true,"in":"header","schema":{"type":"string"}},{"name":"X-API-Key","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":""}},"summary":"Tạo phiên đăng nhập quản trị trong Redis bằng X-API-Key","tags":["Hệ thống"]}},"/api/v1/auth/sessions/current":{"get":{"operationId":"SessionController_current","parameters":[{"name":"authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Xem phiên đăng nhập hiện tại","tags":["Hệ thống"]},"delete":{"operationId":"SessionController_logout","parameters":[{"name":"authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Đăng xuất và thu hồi phiên đăng nhập","tags":["Hệ thống"]}},"/api/v1/deals":{"get":{"operationId":"DealsController_list","parameters":[{"name":"assigned_to","required":false,"in":"query","schema":{}},{"name":"status","required":false,"in":"query","schema":{}}],"responses":{"200":{"description":""}},"summary":"Danh sách giao dịch","tags":["Khách hàng & giao dịch"]}},"/api/v1/config/mappings":{"get":{"operationId":"ConfigController_getMappings","parameters":[],"responses":{"200":{"description":""}},"summary":"Xem ánh xạ trường dữ liệu","tags":["Cấu hình"]},"put":{"operationId":"ConfigController_putMappings","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","example":{"field_mapping":{"lead_data.full_name":"NAME","lead_data.email":"EMAIL[0][VALUE]"}}}}}},"responses":{"200":{"description":""}},"summary":"Cập nhật ánh xạ trường dữ liệu","tags":["Cấu hình"]}},"/api/v1/config/rules":{"get":{"operationId":"ConfigController_getRules","parameters":[],"responses":{"200":{"description":""}},"summary":"Xem quy tắc tạo giao dịch","tags":["Cấu hình"]},"put":{"operationId":"ConfigController_putRules","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","example":{"deal_rules":[{"condition":"campaign.campaign_name CONTAINS 'sale'","action":"create_deal","pipeline_id":"0","stage_id":"NEW","probability":30,"probability_mode":"fixed","assign_to":{"strategy":"round_robin","users":[1,2]}}]}}}}},"responses":{"200":{"description":""}},"summary":"Cập nhật quy tắc tạo giao dịch","tags":["Cấu hình"]}},"/api/v1/config/costs":{"get":{"operationId":"ConfigController_getCosts","parameters":[],"responses":{"200":{"description":""}},"summary":"Xem chi phí chiến dịch","tags":["Cấu hình"]},"put":{"description":"Nhập tổng chi phí VND tương ứng với nhóm lead và khoảng thời gian báo cáo đang đánh giá. Không tự lấy chi phí từ TikTok hoặc quy đổi ngoại tệ.","operationId":"ConfigController_putCosts","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"number"},"example":{"1234567890123456789":5000000}}}}},"responses":{"200":{"description":""}},"summary":"Cập nhật chi phí chiến dịch","tags":["Cấu hình"]}},"/webhooks/tiktok/leads":{"post":{"description":"Trả 202 sau khi đưa vào hàng đợi. Chế độ giả lập: nội dung {}, để trống chữ ký để Swagger tự ký.","operationId":"TikTokWebhookController_receive","parameters":[{"name":"TikTok-Signature","in":"header","description":"Swagger tự ký khi BITRIX24_MOCK=true hoặc MOCK_LEADS_ENABLED=true. Nếu tắt cả hai, phải tự cung cấp chữ ký hợp lệ.","required":false,"schema":{"type":"string"}}],"requestBody":{"required":false,"description":"Payload cần có event và event_id. Trong Swagger ở chế độ giả lập, body {} được interceptor thay bằng lead demo và ký tự động.","content":{"application/json":{"schema":{"oneOf":[{"type":"object","required":["event","event_id"],"properties":{"event":{"type":"string","minLength":1},"event_id":{"type":"string","minLength":1}},"additionalProperties":true},{"type":"object","maxProperties":0,"description":"Chỉ dùng trong Swagger khi bật Bitrix mock hoặc MOCK_LEADS_ENABLED."}]},"examples":{"autoDemo":{"summary":"Giả lập: để {} để tự sinh khách hàng tiềm năng mới và ký yêu cầu","value":{}}}}}},"responses":{"202":{"description":"Sự kiện đã được lưu và đưa vào hàng đợi; sự kiện trùng được xác nhận mà không enqueue lại.","content":{"application/json":{"schema":{"oneOf":[{"type":"object","required":["accepted","id"],"properties":{"accepted":{"type":"boolean","enum":[true]},"id":{"type":"string","format":"uuid"}}},{"type":"object","required":["accepted","duplicate"],"properties":{"accepted":{"type":"boolean","enum":[true]},"duplicate":{"type":"boolean","enum":[true]}}}]}}}},"400":{"description":"Payload thiếu event hoặc event_id dạng chuỗi không rỗng."},"401":{"description":"Thiếu, sai hoặc hết hạn chữ ký. Swagger tự ký khi chế độ tạo dữ liệu TikTok demo được bật."}},"summary":"Nhận khách hàng tiềm năng từ TikTok","tags":["Webhook"]}},"/api/v1/webhook-events":{"get":{"description":"Webhook trả 202 ngay khi đã lưu; dữ liệu sai (email/SĐT không hợp lệ...) chỉ bị phát hiện ở bước xử lý nền và có status=failed cùng trường error.","operationId":"WebhookEventsController_list","parameters":[{"name":"limit","required":false,"in":"query","schema":{}},{"name":"page","required":false,"in":"query","schema":{}},{"name":"event_type","required":false,"in":"query","schema":{}},{"name":"status","required":false,"in":"query","schema":{"enum":["received","processed","failed"],"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Danh sách sự kiện webhook đã nhận (kèm payload gốc và lý do lỗi)","tags":["Webhook"]}},"/api/v1/leads":{"get":{"operationId":"LeadsController_list","parameters":[{"name":"campaign_id","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}},{"name":"source","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"integer","minimum":1,"maximum":100},"description":"Số bản ghi mỗi trang, mặc định 10 và tối đa 100."},{"name":"page","required":false,"in":"query","schema":{"type":"integer","minimum":1},"description":"Trang bắt đầu từ 1."}],"responses":{"200":{"description":"Danh sách lead theo trang và tổng số bản ghi khớp bộ lọc.","content":{"application/json":{"schema":{"type":"object","required":["items","total","page","limit"],"properties":{"items":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"externalId":{"type":"string"},"source":{"type":"string"},"name":{"type":"string"},"email":{"type":"string","nullable":true},"phone":{"type":"string","nullable":true},"city":{"type":"string","nullable":true},"campaignId":{"type":"string","nullable":true},"adId":{"type":"string","nullable":true},"formId":{"type":"string","nullable":true},"extraContacts":{"type":"object","properties":{"emails":{"type":"array","items":{"type":"string"}},"phones":{"type":"array","items":{"type":"string"}}}},"score":{"type":"integer"},"rawData":{"type":"object","nullable":true,"additionalProperties":true},"bitrix24Id":{"type":"integer","nullable":true},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}}},"total":{"type":"integer","minimum":0},"page":{"type":"integer","minimum":1},"limit":{"type":"integer","minimum":1,"maximum":100}}}}}}},"summary":"Danh sách khách hàng tiềm năng","tags":["Khách hàng & giao dịch"]}},"/api/v1/leads/{id}":{"get":{"operationId":"LeadsController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Chi tiết và lịch sử khách hàng tiềm năng","tags":["Khách hàng & giao dịch"]}},"/api/v1/leads/{id}/convert-to-deal":{"post":{"operationId":"LeadsController_convert","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"202":{"description":""}},"summary":"Chuyển khách hàng tiềm năng thành giao dịch","tags":["Khách hàng & giao dịch"]}},"/api/v1/leads/import":{"post":{"operationId":"LeadsController_importBatch","parameters":[],"requestBody":{"required":true,"description":"Mảng sự kiện TikTok. Kết quả từng dòng: queued, duplicate, invalid, failed hoặc pending_recovery.","content":{"application/json":{"schema":{"type":"array","maxItems":1000,"items":{"type":"object","additionalProperties":true}}}}},"responses":{"202":{"description":""}},"summary":"Nhập dữ liệu khách hàng theo lô","tags":["Khách hàng & giao dịch"]}},"/webhooks/bitrix24/deals":{"post":{"operationId":"Bitrix24WebhookController_deal","parameters":[],"requestBody":{"required":true,"description":"Mã xác thực: BITRIX24_APP_TOKEN trong .env. Mã giao dịch: bitrix24Id từ GET /api/v1/deals.","content":{"application/json":{"schema":{"type":"object","required":["auth","data"],"properties":{"event":{"type":"string","example":"ONCRMDEALUPDATE"},"auth":{"type":"object","required":["application_token"],"properties":{"application_token":{"type":"string","example":"REPLACE_WITH_BITRIX24_APP_TOKEN"}}},"data":{"type":"object","required":["FIELDS"],"properties":{"FIELDS":{"type":"object","required":["ID"],"properties":{"ID":{"type":"string","example":"27"}}}}}}}}}},"responses":{"200":{"description":"Đã xử lý hoặc bỏ qua nếu giao dịch không có liên kết trong ứng dụng."},"401":{"description":"Thiếu hoặc sai auth.application_token; phải khớp BITRIX24_APP_TOKEN trên máy chủ."}},"summary":"Nhận cập nhật giao dịch từ Bitrix24","tags":["Webhook"]}},"/api/v1/analytics/stream":{"get":{"description":"Nhập date_range và bấm Bắt đầu bên dưới. Số liệu tự cập nhật khi thay đổi; bấm Dừng để đóng kết nối.","operationId":"AnalyticsController_stream","parameters":[{"name":"date_range","required":false,"in":"query","schema":{"example":"30d"}}],"responses":{"200":{"description":""}},"summary":"Theo dõi số liệu thay đổi theo thời gian thực","tags":["Thống kê & báo cáo"]}},"/api/v1/analytics/conversion-rates":{"get":{"operationId":"AnalyticsController_conversion","parameters":[{"name":"date_range","required":false,"in":"query","schema":{"example":"30d","type":"string"}}],"responses":{"200":{"description":""}},"summary":"Xem tỷ lệ chuyển đổi","tags":["Thống kê & báo cáo"]}},"/api/v1/analytics/campaign-performance":{"get":{"operationId":"AnalyticsController_performance","parameters":[{"name":"date_range","required":false,"in":"query","schema":{"example":"30d","type":"string"}}],"responses":{"200":{"description":""}},"summary":"Xem hiệu quả chiến dịch","tags":["Thống kê & báo cáo"]}},"/api/v1/reports/export":{"get":{"operationId":"AnalyticsController_export","parameters":[{"name":"date_range","required":false,"in":"query","schema":{"example":"30d"}},{"name":"format","required":true,"in":"query","schema":{"enum":["csv","json","xlsx"],"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Tải báo cáo CSV, JSON hoặc Excel","tags":["Thống kê & báo cáo"]}},"/api/v1/reports/deliveries":{"post":{"operationId":"ReportController_send","parameters":[],"responses":{"201":{"description":""}},"summary":"Đưa báo cáo Excel 7 ngày vào hàng đợi gửi đến REPORT_WEBHOOK_URL","tags":["Thống kê & báo cáo"]}},"/api/v1/reports/deliveries/{id}":{"get":{"operationId":"ReportController_status","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Xem trạng thái gửi báo cáo","tags":["Thống kê & báo cáo"]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"summary":"Kiểm tra kết nối PostgreSQL và Redis","tags":["Hệ thống"]}}},"info":{"title":"Khách hàng tiềm năng TikTok → Bitrix24","description":"API tích hợp TikTok → Bitrix24. API quản trị dùng X-API-Key hoặc phiên Bearer; webhook có xác thực riêng.","version":"1.0","contact":{}},"tags":[{"name":"Khách hàng & giao dịch","description":"Danh sách, chi tiết, nhập khách hàng và chuyển thành giao dịch."},{"name":"Thống kê & báo cáo","description":"Chuyển đổi, hiệu quả chiến dịch, số liệu trực tiếp và xuất báo cáo."},{"name":"Cấu hình","description":"Ánh xạ trường, quy tắc tạo giao dịch và chi phí chiến dịch."},{"name":"Webhook","description":"TikTok gửi khách hàng; Bitrix24 gửi cập nhật giao dịch. Tra cứu trạng thái tiếp nhận."},{"name":"Hệ thống","description":"Phiên đăng nhập quản trị và kiểm tra kết nối."}],"servers":[],"components":{"securitySchemes":{"admin":{"type":"apiKey","in":"header","name":"X-API-Key"},"session":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{}},"security":[{"admin":[]},{"session":[]}]}